Hide future entries on Unread page
Some feeds (e.g. [3sat]) contain entries with a publish date
significantly in the future. It's annoying to always see those on the
"Unread" page because the actual content isn't available yet to be
consmed.

Added the `HideFutureEntries` option so users can enable to hide entries
with a future publish date on the "Unread" page.

The signature of `CountUnreadEntries` was changed to accept a
`model.User` to also have access to the value of `HideFutureEntries`.
Passing the information that way seemed cleaner than joining on the
`users` table using the `userID`.

[3sat]: https://www.3sat.de/rss/zdf/kabarett